Add sync and import scripts to upload to NAS
After I've ripped a bunch of discs I use these two scripts to copy them
to my NAS, and import them into my library.

The `sync` script is run on the computer that I've ripped the CD on
(using `ripcd`).

The `import` script is run on my NAS, and uses beets [1] to fetch cover
art and to name files and directories consistently.

[1] See https://beets.io