Don't rsync discs that are currently being ripped
If the `sync` script runs while a disc is being ripped, some of the
tracks will be uploaded and then the folder that FLAC files are being
written to will be moved out of the way. This leads to errors, and at
best the rip needs to be restarted.

After ripping a disc `whipper` creates a `.cue` file for it. So if that
file doesn't exist in a disc's directory we know that the rip hasn't
finished, and we shouldn't try and upload it.