Optionally set genre via environment variable
As soon as a rip finishes and the disc is ejected from the drive, I'm
tempted to run the `upload` script in a separate terminal.

If I don't remember that I've not set the genre yet, freshly ripped FLAC
files can be uploaded to the NAS without having a genre set.

This patch allows the genre to be set on the command line, which means
that the moment the disc is ejected I can safely run the upload script.

Example:

    GENRE=Electronica ./ripcd